Add basic support for full-bleed dialog content.
Doesn't handle full bleed dialogs with actions, assumes certain behavior
of the title bar that matches what we want for the map dialog. Probably
in the future the map preview dialog should be more advanced, and allow
zooming, etc. If we want full bleed content elsewhere, we may need to
generalize this support more, or rearrange how the dialogs are
structured to allow more control.