Use aspect-ratio for sizing maps, reduce some of the reproduced sizing functionality in MapThumbnail.
This stuff is super brittle, so I didn't really *want* to touch it, but
the previous system very much did not work correctly with a lot of
layout primitives (Flexbox was a bit problematic, Grid was very
problematic). MapThumbnail was also very particular to usage in map
browsers but it's really more of a "map image with controls on top"
(will probably rename it soon) so I made the forced aspect ratio a
custom parameter for those, instead of the only config.

I think I caught everything that broke here and fixed it, but it's
possible I missed something. We'll see I guess :)

I think as of this, the mini match history doesn't look great with wide
maps (e.g. Blue Storm), so perhaps we might want to revisit how that
treats things, or provide something that lets us guarantee a <= 1 aspect
ratio or something (so wide maps would be cut off, tall maps would be
shown).