Persist the "join channel" message on the server to the DB. (#758)
This makes it so the "join channel" message is visible even after
reloading the app. Note that we're only persisting the "join channel"
message for now, and not the "leave channel" message. This makes it
possible to a person to join and leave a channel multiple times, and
other users will only see the "join channel" message when they reload,
which might be a bit weird, but doesn't seem like a big deal to me atm.