Highlight text messages if they mention logged-in user.
User is considered mentioned in any text message contains user's name
preceded by the @ character. This applies to chat channels, lobby chat,
party chat, and whispers.

This PR also includes the code for setting the special icon in the
system tray. But first we need to make one. That might happen in this
PR, or at a later point.