Populate the user infos in lobbies. (#767)
Previously, we were relying on user infos being already loaded inside
lobbies, which wasn't really a safe assumption. Creating a lobby from
any page other than the main chat channel could easily lead to someone
joining our lobby for whom we don't have user info loaded.

Now we're loading the user infos at 2 points. When we join the lobby, we
load the user infos for all users that are currently in the lobby, and
when someone new joins the lobby.

We don't clear the user infos when someone leaves the lobby, which might
be a problem for public lobbies? Probably not a big problem.