Add dialogs for policies + require clients to accept the policies when signing up.
This also adds support for error text to checkboxes (although it looks
kinda bad in the signup form atm, but it's such a small + rare surface
area of the app that it's not worth fixing yet imo).

This doesn't yet store policy acceptance in the DB, that'll come in the
next piece of work (when we'll also notify clients of updates).