Add support for connecting to deployed DBs via Tailscale.
The first time this is deployed, you generally need to re-up the
containers twice (once to be able to auth the tailscale instance via:

$ docker exec -it shieldbattery_tailscale_1 /bin/sh
$ tailscale login

Then set the resulting IP for the container in .env as TAILSCALE_IP,
then run `docker-compose up -d` again (to get the db to bind to that
IP). I've had issues in the past where that command needed to be run
twice because tailscale hadn't provided the host interface the first
time, but I think that's not an issue now that the setup is actually
correct, and it should bind it when you log in.