Implement basic support for user mentions in various chat services. (#765)
Also highlights text messages if they mention the logged-in user.

User is considered mentioned in any text message contains user's name
preceded by the @ character, and prefixed/postfixed by separators
(e.g. spaces, most punctuation, beginning/end of message). This applies
to all messaging services (chat channels, lobby chat, party chat, and
whispers). For services that store messages in the DB, mentions are
stored as well.

Parsed mentions are replaced by a special character sequence in
message that we send to clients to make them easy to detect and parse
properly, and to make them based on user ID rather than username, so
they work properly when we add name changes down the road.

This PR also includes the code for setting the special icon in the
system tray (which will be made in a separate commit).

Co-authored-by: Travis Collins <travis@tec27.com>