Fix the mention regex not matching multiple mentions separated by space.
When having two mentions separated by a space, the "postfix" character
of the first mention would interfer with the "prefix" of the second
mention, resulting in second mention not being mentioned.

This was fixed by making the "postfix" group use the positive lookahead
instead of a named capture group, so it doesn't end up in matched
string.

Also added a regression test for this behaviour.