Make channel owners be defined explicitly, instead of implicitly.
Currently, we were using the "edit_permissions" flag as a makeshift
indicator of channel ownership. However, this is not ideal since more
than one person can have this permission set. Technically we could've
defined a channel owner at the model level, where the owner would be set
as a user with this permission and who joined the channel the earliest.

However, I feel like this would get a bit unwieldy and messy down the
line, especially if we start thinking about features like transferring
channel ownership.

This PR adds another column to the database, which ensures that there
can be only one owner per channel, at the database-level. I guess it is
a little bit weird to have a flag that trumps all other flags, no matter
their values, but I believe this is what Discord does too, so we should
be fine? :d