Add proper support for matching 2v2s with the matchmaker.
Matching is done similarly to 1v1 as far as filtering goes, then we
select 3 random players out of the final filtered list, generate all the
possible team configurations for those 4 players, and pick the team
configuration with the lowest difference in effective ratings.

Effective rating is calculated via Root Mean Square, so higher skilled
players are weighted higher in the resulting mean value. For instance,
given a team of a 2500 MMR and 500 MMR player, standard mean gives 1500,
RMS gives ~1800. We can (and probably should) adjust the exponent used
in this calculation once we have some actual data on how much better
skilled players contribute to winrate.