Subscribe to updated game results when viewing the results page for a game that is not yet reconciled.
This should fix cases where clients end up on a results page after a
game before the server can update results (they'll see the older page
and then it should update shortly after). It's also just a nicer
behavior instead of having to click a refresh button to see if it's
changed.

This commit also makes a small change to socket groups such that
subscribing multiple times to the same path is now a no-op instead of an
error. I think originally this was an error because I was worried about
passing different `initialDataGetter` implementations, but this seems
incredibly unlikely with our current patterns (which instead mostly just
retrieve data from services/the DB and are thus the same for any given
path). I also added a boolean when unsubscribing to indicate whether
anything was actually removed, just in case we need to be sure in some
service.