Make game result reporting work properly for games with alliances and players that leave early.
When I originally wrote this code 7 years ago (before it got converted
to Rust), I seemingly thought that the `player_has_left` field
corresponded both to leaving as well as being dropped. Turns out, it
doesn't! So most of the logic here was wrong, and the resulting result
reconcilations were mostly wrong if a player on the winning team left
early (we'd get a win for one player on the team and a loss for the
other).

The code now takes alliances into account, under the assumption that
alliance state cannot be changed from or towards a player once they have
disconnected (they get removed from the dialog, and obviously do not
have a dialog any more themselves). If the player's last state had them
allied + victory with a winning player, and the winning player also had
them allied + victory, they get a victory.

Before we've found a terminal state (that is, before we see a Victory in
the results for any player), players receive a Disconnected status
(before, they'd receive a Defeat).

Ideally I still think these results would be reported from our own
servers so that they're more reliable, but this change at least makes
them behave as the reconciler expects and doesn't seem to worsen any
capability for malicious clients to make them unreconcilable.