Add queueing as a party, up to the point they'd actually enter matchmaking.
Right now this doesn't put any of them into a gameplay activity, but
really that should happen as soon as the race selection process starts.
I left that out because there wouldn't be a great way to unregister them
after, which makes testing annoying. Will add it as part of the next
step (and deal with people in the party being in an activity already).