Remove PartyUser and use SbUserId/SbUser instead.
This is mostly straightforward, and most stuff was fine to just use IDs
(or already did). Major exception being notifications, which is a bit
odd now. I think maybe long-term we should pass relevant info down with
those notifications, but I didn't want to add that processing step to
the notification list on the server right now, so, *shrug*.

For now we just batch request any needed user info for those
notifications and display it once available, which should work okay.
Adding those hooks to the notification popup seems to have exacerbated
some timing issue between the state updates so I worked around that a
bit.