Standardize on Request/Response instead of Payload/Body for http payloads.
We seemingly used the previous forms very haphazardly. For some APIs,
Payload = Request body, Body => Response body. For others, the opposite.
For some APIs, it depended on the particular API call lol.

Anyway, Request/Response is easier to reason about and pick in general,
so hopefully this stays consistent now.