Allow parties to queue into matchmaking, match them against other parties + solos.
This adjusts the matchmaker to use the concept of an "entity" which can
be either a singular player or a party. This is much more complex in
practice! It becomes a lot harder to estimate how many "entities" we
need to fill a particular game. This isn't *that* bad on 2v2 but for 3v3
we'll likely need to rethink the whole match evaluation process as it
will simply be too complex/error-prone.

Lots of changes necessary in here, including things around the
matchmaking state on the client, so that things display properly whether
you were the initial queue initiator or it happened at some later point
in time.

No added tests in the matchmaker test as of this commit, I will try to
add some for party queues in a follow-up (I mostly want to get this in
because it's already ginormous and touching too much stuff).