Add error dialogs when entering matchmaking fails (party or solo). Also includes significant changes to FetchError.
The FetchError changes should make it behave better in general and keep
less resources around (it can now discard the `Response` object
immediately). It also makes handling error codes more easily doable
(before this would have been a real type mess. It still *kind of* is but
it's at least a bit easier).

I think in most cases it'd be better to handle FetchErrors in the places
triggering the http call (rather than the action-creators), but for
matchmaking, it can be called from a bunch of different places that
should handle these errors identically (dialog).