Add updated user profile overlay design. Includes new ConnectedAvatar component.
ConnectedAvatar should make it easier to A) deal with customizable
avatars going forward and B) show a loading state when the user hasn't
been loaded yet, and size/space that loading state the same as the
"real" thing. We still need to figure out how to test these connected
components easily in dev pages, but the ease-of-use concerns outweigh
that here imo.

This also includes a stop-gap way of retrieving account creation date,
this should really just be part of SbUser, but that's too risky/big of a
change for right before this release, so it'll have to wait.