Move whispers init event to a new path to avoid breaking old clients.
This event is not problematic on its own, but previous code changes have
interacted in a weird way here:

1) goToIndex has been broken for the whisper case for some time without
   us realizing, because people have always had a chat channel
2) The chat channel APIs were revved, so old clients connecting to a new
   server don't get a chat channel

Therefore if a user has whisper sessions open on an old client and
connects to the new server: crash. :(

This seemed like the easiest path to fixing it, versus updating the
legacy-only chat APIs to return channels.

This commit also fixes goToIndex to not be broken and moves it to TS to
prevent further breakages.