Fix issue with user overlay popover immediately closing when chat user lists are longer (e.g. in prod).
Fairly nefarious interaction between the various stages of memoization
here:

 - We grab a list of [ID, name] in the channel and sort them by IDs
 - We then want to sort a list of names by name *only* when that first
   list changes

What we were doing here prior to this change was sorting that original
list in place, which is VERY BAD because it means every time that second
memoized function runs, it invalidates the result from the first
function (e.g. that first function will now need to run again next
render, unless the two sorts happen to align). To fix this, we just need
to make a copy of the array before sorting in the second function
(ReadonlyArray helps to enforce this).

There's also some other speculative fixes here that I left in because
they don't seem harmful (e.g. useCallback'ing the expensive selector functions).