Use os.userInfo() to grab current username instead of depending on env vars.
Someone hit a case where process.env.USERNAME was apparently not a
string (likely undefined?), so I'm not sure how consistently available
this is, and this code was (I believe) copied from some example forever
ago. Node has an API to retrieve this info, so we can use that instead,
and just in case it *also* returns undefined, I coalesce it to a default
value as well so we don't crash.