Rewrite all Settings dialog content (minus path dialog) in TS + hooks, add selectedAnnouncer setting.
Fixes #660. Fixes #713.

This doesn't provide total type safety as we call `toJS` on the
immutable values, which basically throws away type info. Ideally we'd
move the settings reducer to immer as well, which would allow us to stop
doing things like that an ensure we are actually using matching
types/keys. May do that in a follow-up now, unsure yet.