Add bans to normal user profile page as well, clean up the ban table structure + API.
I went ahead and deleted the whole admin page because it wasn't
particularly useful right now. Ideally we'll add a user search to the
profile page or something in the very near future and then it won't be
necessary to have some admin-specific page for it. For the time being
you can just find people in the user list in #ShieldBattery, or search
the DB for it.