Greatly simplify the user_ips table and updating.
We were previously storing a ton of data here in a really complicated
fashion that we could never really hope to tease any usefulness out of
(unless we want to spend just as much time writing an IP tracker as we
do the rest of the features :) ). This simplifies the table to better
answer the simple queries: "what IPs has this user used and how often?"
and "what other accounts have used the same IPs?"

In a follow-up commit I'll add this info into the admin user profiles,
and make a job to clear out this day after ~90 days or so.