Add hotkeys to tab items as well.
This PR adds chrome-style hotkeys to all tab items. Mainly Ctrl+1 to
Ctrl+9; with Ctrl+9 being a special hotkey that always focuses the last
tab in the list, regardless of how many tabs there are.

This PR also adds support for all buttons to register multiple hotkeys
for a single button.