Fix results not reconciling when only a losing disconnecter's results are missing.
This is fixed both on the server side and the game side. The game will
not report disconnected players as defeated if there are victorious
players in the game (victorious players = every result should be
terminal). The server will also handle older client's results (with a
disconnect for that player), mapping disconnects -> defeats in results
that have at least one victory.

Still need to write a job to clean up the unreconciled old games, will
do that in the next commit.