Attach RootErrorBoundary twice (between the dialog overlay + the actual app root), check for updates on first launch, fix socket-handlers require path.
We expect the inner RootErrorBoundary to be hit basically always (as
that's where the vast majority of the logic lives), but have the other
one *just in case*. Neither of them have any real info on where the
error came from by their positioning, which is why I just reused the
same error boundary component. Mainly, I want to *maybe* let
auto-updates complete in the event of application errors (I am not
totally certain if this attempt will work on not, given this is hard to
test, will have to figure out a way to do that easily).