Turn off nodeIntegration, add contextIsolation and sandboxing to electron renderer process.
There's probably a bit more we could do to lock down `ipcRenderer`
instead of exposing it directly, but it's kind of hard with the current
webpack setup (and since we can't really generate code from the TS
types). I think this is a pretty good balance of restricted and easy to
maintain going forward. (Also we have CSP which should prevent the vast,
vast majority of potential problems, and ipcRenderer still only exposes
a very limited API)

Fixes #457.