Add SB_RALLY_POINT_PORT env var to control the port rally-point is bound to ingame.
This is mainly useful when debugging network issues (because it can make
Wireshark filters more predictably creatable if you want to watch a
particular client in a multi-client setup). Also useful if you want to
test packet loss on a particular client without affecting the others.

Definitely not recommended to use this setting outside of that, as it
can cause multiple running clients to conflict with each other and fail
to launch properly.