Add an end-to-end test for signing up and verifying email.
This includes an upgrade to our mailgun package (to the official
library), an added env flag to send mailgun requests to a particular URL
(meant for use in integration tests), and a small fake server for
handling mailgun requests and logging sent emails, which is now spun up
and used by integration tests.

I also discovered that the changelog showing logic was... very broken so
I fixed that so that I could suppress it in the new test.