Build infrastructure for message duplication + ack management.
This largely builds on the system described here:
https://gafferongames.com/post/reliable_ordered_messages/

I did, however, make a number of tweaks given that we're using protobuf
(free varint implementation), and I've seen some more clever things in
QUIC that we could probably use in the future. For the moment these
tweaks probably make our overhead *slightly* larger than the original
implementation, especially after 20-30 minutes of play, but we're
talking about a few bytes. There's more overhead to be removed from
getting Storm out of the netcode process in the near future, anyway.

This infrastructure is not yet used in actually processing packets, I'll
do that in the next commit.