Adjust max latency calculating and static turn rate logic to better match reality.
I did a lot of thinking, and diagramming, and testing, and realized that
Grant's post about this is actually more wrong than I thought. This ends
up being the source of all (well, most) of the problems we previously
saw with static turn rate calculation.

Basically, Grant's latency formula assumes that players know how to time
travel, which, as far as I am aware, they do not (maybe they do and
that's why I'm bad at BW :) ). His formula is:

`1000 * (UserDelay + 1) / TR`

For this to be the case, users would have to be able to send their
current turn before they have actually played it, which is clearly
unworkable. So the *actual* formula is `1000 * UserDelay / TR`. *Except*
BW also appears to be extremely inconsistent in how quickly it runs
turns. Because of this, an individual player's turn might be ~14ms late,
while their opponent's is ~14ms early, and this happens *frequently*. So
basically, there's a static overhead factor as well that is a large % of
the normal turn time.

So in short, our new formula for max latency is:

`1000 * UserDelay / TR - 28`

The constant overhead factor could probably use more sampling (e.g. if
14ms turns out to not be hit 99% of the time, we could probably adjust
it down), but I think being safe with static turnrates is probably
better now anyway. If we can't improve BW's turn timing, then at the
very least the next rev of netcode can track individual player's turn
timings and factor it into the latency choice.