Replace SC:R's GetTickCount calls with a high resolution timer.
This makes turns run at a *far* more consistent rate, which should
improve netcode reliability and allows us to drop a few ms off our
network overhead in the max latency calculation. I would think the
effect would actually be larger than it seems to be, but I guess
probability-wise there are times where delayed/fast turns line up
anyway? In any case, when we have a system that can change delay
dynamically we won't have to be so conservative with this estimate.