Fix issues with packet loss + players leaving.
Prior to this change, we blocked all resend requests because they
weren't generally useful. *But* in the very specific case of a player
leaving the game with packet loss, these become necessary. Consider:

4 players: A, B, C, D. A has packet loss. Players play turn N and send
  each other packets for it, all of A's get lost except for the one to
B. B can now simulate up to turn N, while C and D can only simulate up
to N-1. A leaves immediately after turn N (no further packets
delivered). All players see a drop screen, one of them chooses to drop
A. At this point, C and D no longer need to wait for the turn N packet
from A (they're disconnected) so they simulate turn N. But their turn N
is now different from B's (it didn't have A's input), so they see this
as a desync and drop B.

Resend requests "fix" this (kind of haphazardly) by allowing B to
communicate the packets from A during the stall/drop screen.

We don't want to allow *all* resend requests, because they are still
mostly useless with our new protocol, so the new check simply looks to
see if we've received a packet from the target player in the last 500ms.
If we have, we assume Storm is only missing a single message and it will
catch up shortly. Otherwise, we allow the request through.

This commit also adds additional attempts to deliver unacked payloads
before the game process quits. After the game finishes we begin sending
empty payloads to any other still-playing clients that we have unacked
payloads for. This triggers AckManager to fill those packets with the
unacked payloads, and (hopefully) get acks for them. If these payloads
are e.g. the "I'm leaving the game" command, then this helps prevent the
other players from seeing a drop screen (or needing resends).