Add an API to get and update user's permissions in a chat channel.
There's a bit of weirdness with updating the `owner` permission which I
disabled for now, as I believe a dedicated API to transfer channel
ownershipt might be more appropriate. But this leads me to believe that
this might not be the best place to keep the information about who's the
channel owner. Instead, it might make more sense to keep that
information in the `channel` table where each row can have a foreign key
to the `user` table, referencing the channel owner. This should help
make it certain that we don't end up in a weird state, like having
multiple owners per channel.