Add race statistics tracking for ladder games.
So far we only tracked race statistics for all games a user has played
in aggregate. Now we're tracking them for only their ladder games as
well, partitioned by each matchmaking type.

Note that we're currently not displaying this information anywhere on
the client. That will come in a separate PR.