Upgrade to React 18 (also includes a bunch of other minor dep updates because I don't know how to make yarn upgrade transitive deps without doing that :( ).
Fixes are mostly due to:

1) Needed to move to a more modern hot reloading system (React-Refresh)
   since react-hot-loader isn't updated for 18
2) Needed to move to React.lazy from @loadable/component. React.lazy has
   a less nice API that expects only default exports, probably we could
   write a little wrapper around it to clean that stuff up a bit
3) A bunch of typings needed fixing since React is a bit more strict now

Only behavioral thing I saw was on the ladder page, which appears to be
a bug introduced in the Virtuoso change rather than here, but I didn't
bisect it to be sure.

No new React features used yet (outside of the new render call), those
can be added in future commits.