Add support for matchmaking seasons.
* Add matchmaking_seasons table + columns in matchmaking_ratings tables.

Seasons have a start time and a name, this assumes we run the same
season for every matchmaking type, and that there is no break period
between seasons. I think this seems fine, if we need the more complex
behavior we can always add it in the future.

Currently the seasons table also has a flag for using "legacy ratings",
which will make the matchmaking code use the old rating calculation.
The plan here is that we release a version with this prior to the next
season start with both the old rating calculation and new, set the next
season start time, and then after the next season starts we can remove
this column + the legacy rating code. Mainly I didn't want to have the
standard developer setting involve having to set this flag to false
manually on the first season, so this just allows an easy way to migrate
over a couple of releases.

* Add a service for retrieving + caching matchmaking season info.

* Only rank users from the latest season in ranked view (archival ranks can be placed in a static table, but that's future work not yet done here).

* Don't update MMRs/rankings when reconciling games for a past season.