Wait 1 second before reporting unhandled rejections.
The `unhandledRejection` event happens somewhat early/unpredictably, and
in some cases we just haven't yet been given a chance to add an error
handler (but will). To deal with this, we can use `rejectionHandled` to
remove erroneous cases, but we still need to pick some limit to report
these things at. 1 second seems reasonable to me at the moment, but if
this causes problems we can re-evaluate.

Fixes #847.