Keep message input contents across page changes in chat/whispers.
This should be done for lobbies + parties as well, but it's a bit more
complex there, as we don't really have a good identifier to utilize for
the key right now. Perhaps we could generate an ID (on the client) when
either of these are first joined and use that? [Mainly we want to avoid
restoring state if you leave one party and join another, or if you leave
a lobby and then join another with the same name, etc.]

We currently store these in sessionStorage, which we now clear whenever
auth would change (signup, login, logout). This is the only thing we put
in sessionStorage right now so this seemed safe to change, and makes it
so we can assume sessionStorage is always relative to a particular
account, which seems nice.