Implement a post-match results dialog (for matchmaking only).
This will pop up only if the results are for the last game that was
played, regardless of whether you're still looking at the results page.
No fancy transitions/sounds are implemented yet, that'll come in a
follow-up commit. The replay button also doesn't work as of this commit
(it just shows a snackbar), but that will be the next thing I work on.

Other things included in this commit:

  - Fixes for dialog payload typings so payloads can actually be
    required
  - Binary search implementation for arrays
  - Removed max player limitation in Champion rank, which didn't really
    work very well in our system (now anyone that is 2400+ will be
    placed in Champion)