Don't go to active-game page when a replay starts.
I'm not really sure when this got added (was it always there?) but there
was zero point in sending people there, as it will just immediately send
them back to the index (there is no "active game" as far as the client
is concerned at that point). If we have some replay-companion-UI at some
point maybe it would make sense to send them there, but as is I don't
see a point in doing any kind of navigation for this.