Rework channel joining.
Includes a new channel-status-card component that is for now only used
to display the info for a channel that user hasn't joined yet, but will
probably be used in other places too, like mentioning the channel in
chat.

Also has a few changes that were broken in a previous commit.