Add initial support for friending and blocking users.
This uses a database structure that allows most queries to be done in a
single update step, because the rows track the mutual state between
users. This also makes it easier to enforce the necessary constraints
around this:

- Users can friend or block someone, but not both
- Blocked users cannot receive friend requests from users they've
  blocked
- Friendship requires mutual acceptance

To accomplish this, state is stored in the DB with user IDs sorted (low
and high), and we add a few states to define the direction of
relationships that can exist non-mutually (friend requests, blocks). All
operations can then be done in a single step (with the exception of
removing blocks, which can potentially be either an update or a delete
and thus require us to do a `SELECT ... FOR UPDATE`).

This change implements the UI for displaying friends as well as
incoming/outgoing requests, and adds various actions to the user context
menu. A number of things are not yet implemented here, and will be done
in follow-up commits (some may also happen in later versions):

- Buttons to accept/reject requests in the friends popover
- Tracking online/offline state of friends
- Effects of blocks outside of friend requests (blocking messages,
  whispers, etc.)
- UI for managing blocks in settings
- Search box in the friend popover for filtering friends + finding users
  to add as a friend