Switch to stimpack injector
Stimpack has slightly different api; as it combines launch and inject
steps to one function, as well as automatically terminates the process
on launch failures.

Game side had to slightly change too: A new sb_init.dll is needed to be
compatible with the new injection method, and the new dll now passes
us control (just a tiny bit) after process entry. So we don't have to hook
the entry point anymore, and can immediately start the async thread.

# Conflicts:
#	app/game/active-game-manager.ts