Rename the "high_traffic" column to "official" in the channels table.
Official channels have the following characteristics:
- they're created by staff members only
- they don't have owners (so only server moderators can moderate them)
- they're not deleted when everyone leaves (thus keeping the message
  history in perpetuity)
- TODO MAYBE: join/leave messages are not shown in them
- TODO MAYBE: new accounts are automatically joined in all (some?) of
  them

TODO: Currently, every channel that is created is a non-official channel
(except ShieldBattery which was created as an official channel in a
migration), so once we make it possible to update channel settings, we
gotta make sure to set the channel owner to NULL if the channel is being
made into an official channel. I don't think it's that big of a deal
even if we don't, since only staff will most likely be able to make
official channels, but it'll make things easier in case of de-modding or
something.