Fix using context menus inside popovers causing both the popover + context menu to close.
This adds a PortalContext that manages the layering for Portals, such
that only the root Portal attaches click handlers and delegates those
events to the nested Portals. By doing so, we can check if any of the
nested Portals contain the click event (which is necessary because each
Portal is a sibling at the root document level, rather than being nested
like the React component tree).

I think a cleaner way to handle this in the future (without
reimplementing event delegation like this) might be to use the
`<dialog>` element (which allows for this top-level kind of UI without
requiring a portal). Instead of handling click events, we could utilize
focus events (I *think*) to determine when clicks outside the top-level
UI occur without requiring a document handler, and allow events to
follow the normal document order. This change would be a lot more
significant and have much more potential to cause bugs, so I didn't
pursue it here.

As of this change there's still some potentially undesirable behavior
with nested Portals, namely that clicking outside of all of them will
dismiss them all at once. I *think* a more desirable behavior would be
to only dismiss the top one. This could be handled with a change to the
capture handlers (non-top-level Portals could always be told that the
event was contained?) but I left that out for now to see how it feels
for a day or two.