Handle blocks in whispers + party invites.
In whispers, we filter client-side. Blocked users are allowed to send
whispers, but they are not displayed to the blocker in-client. If they
were to unblock, they would then see these messages.

In parties, we do the check server-side, blocked users are not allowed
to send invites to the blocker (and are told as such). Parties may
contain a user that is blocked by another party member, such as if a 3rd
leader invites both users. In this case it will basically behave like
a chat channel (messages from the blocked user will be hidden for the
blocker, and not trigger notifications).

Also includes a fix for the timestamp on blocks created on top of other
relationships (e.g. friend request -> block).