Remove custom title from ShieldBattery BW window.
This is a very speculative fix for people reporting mouse sensitivity
differences between SB and Bnet. We don't see a lot of reasons for the
behavior they report, but I know some of them use Razer mice, and I do
not have high hopes that Razer software isn't doing something weird
based on Process name + Window name to adjust mousing. No idea that this
will actually fix anything, but reduces the likelihood of customizations
for a Bnet not applying to our BW instances as well, so, we'll see.