Game: Forge: Replace GetWindowLong hook with SetWindowLong
Instead of having a hook in GetWindowLong and always add the
WS_VISIBLE flag in case the GWL_STYLE is queried.

Hook SetWindowLong instead and fixup the flags being passed prior
to calling the original SetWindowLong in case it's the SC:R main window.

This fix is required to get Wine working properly, otherwise the Window
is left in a hidden state. (Probably because Wine code calls GetWindowLong
as well, and the hook is a global trampoline on the DLL code).

Signed-off-by: Jon Doron <arilou@gmail.com>