Allow logged out users to view leagues, ladder, and user profiles.
I think it would be good to let them view games as well, but I think
that page/API has some assumptions around websocket connectivity, so I
didn't try to deal with it here (will file an issue for this).

It'd probably be nice to share a bit more around the routes + activity
bar stuff between logged out/logged in (and maybe just rethink the
top-level structure more generally).