Only use pretty-id for client-side routing, convert it into normal UUIDs everywhere else.
After using the previous form (APIs take in pretty-id form and convert
them to UUID there), I encountered a ton of pain in converting the data
we give to clients since that all needed to have pretty-ids as well, or
they wouldn't be able to match things up.

Using these nicer IDs for routing only provides basically all the
benefits (humans see them), while not requiring us to do any conversion
of the underlying data when giving it to clients, so it seems like a
much better approach.