Add a script for pre-downloading Google Fonts, use self-hosted URLs in app/web.
This doesn't happen automatically, we'll have to re-run the script
manually if the fonts change (but this is easy to do). We could pretty
easily add it into the deploy script or something, but I haven't
bothered yet.

The fonts for web go into the public directory, so they should get
pushed to the CDN as well.

Note that the utility for this is ignored by TS currently because one of
the font downloader's dependencies fails to typecheck.

Fixes #891.