Use UUIDv7 instead of v6.
v7 is more similar to v4 (which we were using before), whereas v6 is
similar to v1 (which contains a machine identifier that isn't
particularly useful/desirable to us). I think this function should also
perform slightly better given that it can use the underlying v4
infrastructure for generation. This change is again compatible with any
previously generated UUIDs.